requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==1.0.0
marshmallow==3.21.1
python-magic-bin==0.4.14
flask-cors==4.0.0
pytest==8.0.0
pytest-cov==4.1.0
python-dotenv==1.0.1
Werkzeug==3.0.1
